fix: code review findings — overflow protection, clear fix, tests
F3 (u16 overflow): clamp u64→u16 at conversion sites with min(u16::MAX)
F4 (clear doesn't reset): clear() now adds cleared lines to evicted_lines
F6 (unseen overflow): widen unseen_lines/total_at_follow to u32
F7 (test gaps): add multi_line_system_message, clear_adds_to_evicted_lines

F1/F2 (Tool/Assistant line count) are mock-llm only (off by default).
F5 (frame mismatch) is bounded by the 100ms tick interval (~1 line).

+++ src/ui/render_state.rs
@@ -33,7 +33,7 @@ pub struct RenderState {
     pub follow_tail: bool,
     #[allow(dead_code)]
     pub prev_total_lines: u16,
-    pub unseen_lines: u16,
+    pub unseen_lines: u32,
     pub evicted_lines: u64,
     pub buffer_total_lines: u64,
     pub panel_visible: bool,
